requirements.txt
numpy>=1.21.0
langchain>=0.2.0
langchain-openai>=0.1.0
python-dotenv>=1.0.0
